+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/KT.kt
@@ -33,7 +33,8 @@ enum class Case(val file: String) {
 	NamingConventions("/cases/NamingConventions.kt"),
 	NewLineAtEndOfFile("/cases/NewLineAtEndOfFile.kt"),
 	MaxLineLength("/cases/MaxLineLength.kt"),
-	MethodNameEqualsClassName("/cases/MethodNameEqualsClassName.kt"),
+	MethodNameEqualsClassNameNegative("/cases/MethodNameEqualsClassNameNegative.kt"),
+	MethodNameEqualsClassNamePositive("/cases/MethodNameEqualsClassNamePositive.kt"),
 	ModifierOrder("/cases/ModifierOrder.kt"),
 	OverloadedMethods("/cases/OverloadedMethods.kt"),
 	ComplexClass("/cases/ComplexClass.kt"),
