+++ detekt-rules/src/test/resources/cases/MethodNameEqualsClassNamePositive.kt
@@ -1,0 +1,34 @@
+@file:Suppress("unused", "FunctionName")
+
+package cases
+
+class MethodNameEqualsClassName {
+
+	fun methodNameEqualsClassName() {} // reports 1
+}
+
+object MethodNameEqualsObjectName {
+
+	fun MethodNameEqualsObjectName() {} // reports 1
+}
+
+class StaticMethodNameEqualsClassName {
+
+	companion object {
+		fun StaticMethodNameEqualsClassName() {} // reports 1
+	}
+}
+
+class MethodNameContainer {
+
+	class MethodNameEqualsNestedClassName {
+
+		fun MethodNameEqualsNestedClassName() {} // reports 1
+	}
+}
+
+class AbstractMethodNameEqualsClassName : BaseClassForMethodNameEqualsClassName() {
+
+	// reports if overridden functions are not ignored
+	override fun AbstractMethodNameEqualsClassName() {}
+}
